network_stats separation and click usage
