+++ ray_on_golem/__main__.py
@@ -1,0 +1,3 @@
+from ray_on_golem.main import main
+
+main()
