+++ ray_on_golem/network_stats/__init__.py
@@ -1,0 +1,3 @@
+from ray_on_golem.network_stats.main import main
+
+__all__ = ("main",)
