+++ ray_on_golem/server/__init__.py
@@ -1,0 +1,3 @@
+from ray_on_golem.server.main import main
+
+__all__ = ("main",)
